+++ backend/src/index.ts
@@ -2,8 +2,8 @@ import "dotenv/config";
 import express from "express";
 import cors from "cors";
 import bodyParser from "body-parser";
-import canvasEditorRouter from "./modules/canvasEditor";
-import questionModuleRouter from "./modules/informationTabs/questionModule";
+import canvasEditorRouter from "./features/validationServices";
+import questionModuleRouter from "./features/informationTabs/questionTab";
 
 const app = express();
 const port = 3001;
